Implement commandline arguments
