+++ src/main.rs
@@ -1,8 +1,14 @@
-use pdf_summarizer::*;
+use clap::Parser;
+use pdf_summarizer::args::{Args, Commands};
+use pdf_summarizer::ollama_api;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let path = std::env::args().nth(1).expect("'path' must be provided.");
-    summarize(std::path::Path::new(&path)).await?;
+    let Args { model, command } = Args::parse();
+    match command {
+        Commands::Summarize { path } => ollama_api::summarize(&path, &model).await?,
+        Commands::Name { path } => ollama_api::name(&path, &model).await?,
+        Commands::Chat => ollama_api::chat().await?,
+    }
     Ok(())
 }
